requirements.txt
django-colorfield
django
django-livereload